refactor: modularize managers
