+++ js/managers/index.js
@@ -1,0 +1,5 @@
+export { BattleLogManager } from './BattleLogManager.js';
+export { VisualEffectManager } from './VisualEffectManager.js';
+export { EventManager } from './EventManager.js';
+export { StatusEffectManager } from './StatusEffectManager.js';
+export { battleMaster } from './BattleMaster.js';
